+++ prozess/fluidfactorprocess.cpp
@@ -69,7 +69,7 @@ ProjectProcess::ResultCode FluidFactorProcess::init( const QMap<QString, QString
     }
 
     Grid2DBounds bounds=m_intercept->bounds();
-    m_grid=std::shared_ptr<Grid2D<double> >( new Grid2D<double>(bounds));
+    m_grid=std::shared_ptr<Grid2D<float> >( new Grid2D<float>(bounds));
 
 
     if( !m_grid){
