+++ prozess/horizonsemblanceprocess.h
@@ -32,8 +32,8 @@ private:
     size_t m_halfSamples;
 
     std::shared_ptr<SeismicDatasetReader> m_reader;
-    std::shared_ptr<Grid2D<double> > m_horizon;
-    std::shared_ptr<Grid2D<double> > m_slice;
+    std::shared_ptr<Grid2D<float> > m_horizon;
+    std::shared_ptr<Grid2D<float> > m_slice;
 
     std::shared_ptr< Array2D<seismic::Trace*> > traceBuffer=nullptr;
 };
